簡略化したfaddをfsimに追加。to_simの更新。

+++ to_sim/floor.c
@@ -4,5 +4,5 @@
 #include "floor.h"
 
 float FloorFloat(float f){
-  return (float)floor(f);
+  return floorf(f);
 }
